app/requirements.txt
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.2
decorator==5.1.1
Django==4.2.3
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
executing==1.2.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
ipython==8.14.0
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
pluggy==1.2.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pure-eval==0.2.2
pycparser==2.21
pycryptodome==3.18.0
Pygments==2.15.1
PyJWT==2.8.0
pytest==7.4.0
pytest-django==4.5.2
python-creditcard @ git+https://github.com/maistodos/python-creditcard.git@6b6bcc5d975985337c35e5c1a4f87d861bcac731
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
wcwidth==0.2.6
